requirements.txt
django-registration>=2.2
django-crispy-forms>=1.6.1
django-recaptcha>=1.2.0
python-rtkit>=0.7.0
requests>=2.12
lxml>=3.7.2
Pillow>=4.0.0
gunicorn==19.9.0
psycopg2==2.7.5
django-multiselectfield==0.1.8
